working start on fakenil.py
fakenil.py is now working as expected. it will run the cross given, and
print a summary of each chromsome (and sister, so 2*n lines). Each line
is a list with the 0th equal to the starting parent, then each
subsequent value is the location of a crossover. Finally the last line
is a proportion of total genome from parent 1.

Todo now is stretch this out to the coordinates of input genomes, and
create a "fake NIL" from the parental genomes